news_automation/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai==0.28
python-dotenv==1.0.0
psycopg2==2.9.10